refactor(csr): merge plasticity_binary_csc into plasticity_binary
Combine the CSC plasticity wrappers (update_csc_on_binary_pre/post) into
plasticity_binary.py alongside the CSR primitives they delegate to, and
merge the corresponding test modules. Update _csr package and main.py
imports to the unified module.
