+++ brainevent/_csr/__init__.py
@@ -19,11 +19,9 @@ from .binary_indexed import binary_csrmv_indexed, binary_csrmv_indexed_p
 from .binary_indexed_mm import binary_csrmm_indexed, binary_csrmm_indexed_p
 from .float import csrmv, csrmv_p, csrmm, csrmm_p
 from .main import CSR, CSC
-from .plasticity_binary_csr import (
+from .plasticity_binary import (
     update_csr_on_binary_pre, update_csr_on_binary_pre_p,
     update_csr_on_binary_post, update_csr_on_binary_post_p,
-)
-from .plasticity_binary_csc import (
     update_csc_on_binary_pre,
     update_csc_on_binary_post,
 )
