+++ brainevent/_csr/plasticity_binary_test.py
@@ -23,7 +23,8 @@ import pytest
 import scipy.sparse as sp
 
 import brainevent
-from brainevent._csr.plasticity_binary_csr import (
+import brainevent as be
+from brainevent._csr.plasticity_binary import (
     update_csr_on_binary_pre,
     update_csr_on_binary_post,
     update_csr_on_binary_pre_p,
@@ -265,3 +266,83 @@ class Test_on_post:
         assert jnp.allclose(csr.todense(), mat, rtol=1e-1, atol=1e-1)
 
         jax.block_until_ready((mat, post_spike, pre_trace))
+
+
+# ---------------------------------------------------------------------------
+# CSC-layout plasticity updates (``update_csc_on_binary_pre`` /
+# ``update_csc_on_binary_post``), validated against a dense reference.
+# ---------------------------------------------------------------------------
+
+
+def _dense(rng, n_pre, n_post, p=0.5):
+    mask = rng.random((n_pre, n_post)) < p
+    vals = rng.random((n_pre, n_post)) + 0.5  # in [0.5, 1.5)
+    return jnp.asarray(mask * vals, dtype=jnp.float32)
+
+
+def _ref_pre(W, pre_spike, post_trace, w_min, w_max):
+    W = np.asarray(W)
+    mask = (W != 0.0)
+    active = (np.asarray(pre_spike) != 0).astype(W.dtype)        # (n_pre,)
+    delta = mask * active[:, None] * np.asarray(post_trace)[None, :]
+    return np.clip(W + delta, w_min, w_max)
+
+
+def _ref_post(W, pre_trace, post_spike, w_min, w_max):
+    W = np.asarray(W)
+    mask = (W != 0.0)
+    active = (np.asarray(post_spike) != 0).astype(W.dtype)       # (n_post,)
+    delta = mask * np.asarray(pre_trace)[:, None] * active[None, :]
+    return np.clip(W + delta, w_min, w_max)
+
+
+@pytest.mark.parametrize("spike_dtype", [jnp.bool_, jnp.float32])
+def test_update_csc_on_binary_pre_matches_dense(spike_dtype):
+    rng = np.random.default_rng(0)
+    n_pre, n_post = 4, 6
+    w_min, w_max = 0.0, 1.2
+    W = _dense(rng, n_pre, n_post)
+    csc = be.CSC.fromdense(W)
+    pre_spike = jnp.asarray(rng.random(n_pre) > 0.5, dtype=spike_dtype)
+    post_trace = jnp.asarray(rng.random(n_post), dtype=jnp.float32)
+    new_w = be.update_csc_on_binary_pre(
+        csc.data, csc.indices, csc.indptr, pre_spike, post_trace,
+        w_min, w_max, shape=csc.shape,
+    )
+    got = csc.with_data(new_w).todense()
+    ref = _ref_pre(W, pre_spike, post_trace, w_min, w_max)
+    assert jnp.allclose(got, jnp.asarray(ref), atol=1e-5)
+
+
+@pytest.mark.parametrize("spike_dtype", [jnp.bool_, jnp.float32])
+def test_update_csc_on_binary_post_matches_dense(spike_dtype):
+    rng = np.random.default_rng(1)
+    n_pre, n_post = 5, 3
+    w_min, w_max = 0.0, 1.2
+    W = _dense(rng, n_pre, n_post)
+    csc = be.CSC.fromdense(W)
+    pre_trace = jnp.asarray(rng.random(n_pre), dtype=jnp.float32)
+    post_spike = jnp.asarray(rng.random(n_post) > 0.5, dtype=spike_dtype)
+    new_w = be.update_csc_on_binary_post(
+        csc.data, csc.indices, csc.indptr, pre_trace, post_spike,
+        w_min, w_max, shape=csc.shape,
+    )
+    got = csc.with_data(new_w).todense()
+    ref = _ref_post(W, pre_trace, post_spike, w_min, w_max)
+    assert jnp.allclose(got, jnp.asarray(ref), atol=1e-5)
+
+
+def test_update_csc_on_binary_pre_no_clip():
+    # Without bounds the rule is a pure additive update at stored positions.
+    rng = np.random.default_rng(2)
+    n_pre, n_post = 3, 4
+    W = _dense(rng, n_pre, n_post)
+    csc = be.CSC.fromdense(W)
+    pre_spike = jnp.asarray([True, False, True])
+    post_trace = jnp.asarray(rng.random(n_post), dtype=jnp.float32)
+    new_w = be.update_csc_on_binary_pre(
+        csc.data, csc.indices, csc.indptr, pre_spike, post_trace, shape=csc.shape,
+    )
+    got = csc.with_data(new_w).todense()
+    ref = _ref_pre(W, pre_spike, post_trace, None, None)
+    assert jnp.allclose(got, jnp.asarray(ref), atol=1e-5)
